requirements.txt
requests==2.26.0
boto3==1.18.65
numpy==1.20.3
pandas==1.3.4
backoff==2.2.1
urllib3==1.26.14
pyjwt==1.7.1
aiohttp==3.8.4
async-timeout==4.0.2
asynctest==0.13.0
aws-mfa==0.0.12
kafka-python==2.0.2